Refactor parser class
