+++ src/main/java/seedu/duke/parsermanager/ParseExit.java
@@ -1,0 +1,31 @@
+package seedu.duke.parsermanager;
+
+import seedu.duke.command.Command;
+import seedu.duke.command.CommandBye;
+import seedu.duke.exception.ExtraParametersException;
+
+import static seedu.duke.Messages.MESSAGE_BYE_PARAMETERS_PRESENT;
+
+public class ParseExit extends Parser {
+
+    private final String commandDetail;
+
+    public ParseExit(String exitCommandDescription) {
+        this.commandDetail = exitCommandDescription;
+    }
+
+    @Override
+    public Command parseCommand() throws ExtraParametersException {
+        checkForEmptyDescription(commandDetail);
+        return new CommandBye();
+    }
+
+    private void checkForEmptyDescription(String commandDetail) throws ExtraParametersException {
+        boolean isNotEmpty = !commandDetail.trim().isEmpty();
+
+        if (isNotEmpty) {
+            throw new ExtraParametersException(MESSAGE_BYE_PARAMETERS_PRESENT);
+        }
+
+    }
+}
